implement configuration POST, implement dynamic querying by Serenity package, folder restructures

+++ src/PlasticisingTile.API/DTO/Plasticising/PlasticisingSerieDto.cs
@@ -2,8 +2,8 @@
 
 namespace PlasticisingTile.API.DTO.Plasticising;
 
-public class PlasticisingSerieDto : ISerieDto<int>
+public class PlasticisingSerieDto : ISerieDto<double>
 {
     public string Name { get; set; } = string.Empty;
-    public IEnumerable<int> DataPoints { get; set; } = new List<int>();
+    public IEnumerable<double> DataPoints { get; set; } = new List<double>();
 }
